+++ concrete_datastore/concrete/constants.py
@@ -1,4 +1,5 @@
 # coding: utf-8
+from decimal import Decimal
 
 ATLEAST_LEVEL_ATTRS = {
     'simpleuser': {'is_active': True},
@@ -62,4 +63,15 @@ TYPE_EQ = {
     'DateTimeField': 'datetime',
     'DateField': 'date',
     'BigIntegerField': 'int',
+}
+
+EMPTY_VALUES_MAP = {
+    'TextField': {'empty_value': '', 'field_type': str},
+    'CharField': {'empty_value': '', 'field_type': str},
+    'IntegerField': {'empty_value': 0, 'field_type': int},
+    'FloatField': {'empty_value': 0.0, 'field_type': float},
+    'DecimalField': {'empty_value': Decimal(0), 'field_type': Decimal},
+    'JSONField': {'empty_value': {}, 'field_type': dict},
+    'BigIntegerField': {'empty_value': 0, 'field_type': int},
+    'BooleanField': {'empty_value': None, 'field_type': bool},
 }